Inline smart cdn method
